add registration and login add error, common pages

+++ src/main/java/com/my/persistence/dao/BaseDAO.java
@@ -9,6 +9,6 @@ public interface BaseDAO <E extends AbstractBaseEntity>{
     E create(E entity);
     E update(E entity);
     boolean delete(String id);
-    Optional<E> findById(Integer id);
+    Optional<E> findById(Long id);
     Collection<E> findAll();
 }
